add option for ocv-soc table

+++ interconnect/genix/genix_op.h
@@ -14,7 +14,7 @@
 
 typedef struct _GENIX_DATA
 {
-	uint16_t 	pack_soc;
+	uint8_t 	pack_soc;
 	uint16_t	pack_voltage;
 	uint16_t	pack_max_temp;
 	int16_t		pack_current;
